requirements.txt
fastapi
uvicorn
opencv-python
ultralytics
paho-mqtt
